requirements.txt
torch
torchvision
ftfy
regex
tqdm
Pillow
transformers
pymilvus
scikit-learn
matplotlib
seaborn